+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add `supportDownlinkBandwidthEstimation` API to check whether the browser support downlink bandwidth estimation 
+  which requires for priority based downlink policy to work.
+
+### Removed
+
+### Fixed
+
+### Changed
+
 ## [2.24.0] - 2021-12-07
 ### Added
 - Add error name for custom device controller error.